requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pytesseract
pillow
